Fix #6: Unconditional modifications to CMSimple_XH are bad
The unconditional `XH_isAccessProtection`modification is really bad,
because after uploading this CMSimple_XH mod to a public webserver the
access protection check would always yield success while the files
might actually be accessible.

We fix that by explicitly checking for the built-in webserver and the
expected `SERVER_NAME`, and only skipping the actual check in this
case.